streamlit_app/requirements.txt
streamlit
openai
pytesseract
pillow
lasio
pandas
PyPDF2
python-docx
